server/pyproject.toml
[tool.poetry]
name = "bagchal-server"
version = "0.1.0"
description = ""
authors = ["Daze <dazehere@yandex.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.1"
alembic = "^1.5.7"
pydantic = "^1.8.1"
psycopg2-binary = "^2.8.6"
SQLAlchemy = "^1.4.2"
pytest = "^6.2.2"
uvicorn = {version = "^0.17.6", extras = ["standard"]}
click = "^7.1.2"
python-dotenv = "^0.17.0"
aiofiles = "^0.6.0"
redis = "^4.3.4"
hiredis = "^2.0.0"
PyYAML = "^6.0"
numpy = "^1.23.1"
tensorflow-macos = "^2.9.2"
pandas = "^1.4.3"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
pylint = "^2.12.2"
rope = "^0.22.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
